This dataset was obtained by the processing of the Establishments table that contains all the data about the active establishments in the Brazil, including the main activity and address of these establishments. The full table can be obtained by the link: https://dados.gov.br/dados/conjuntos-dados/cadastro - national-of-the-legal-person---cnpj .
From this database, the Geopandas library was used to obtain the spatial location of these establishments (lat/long) using the Open Street Maps Nominatim service.

The file available in this repository contains information about health, education, business and leisure establishments. A detailed description of the activities contained in each of these categories is available on the Read.me file